requirements.txt
google-adk>=0.1.0
google-generativeai>=0.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
